refactor: moving gameplay to separate file
